# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NetSage API
  version: 0.1.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
paths:
  /api/auth/register:
    post:
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: JWT token
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: JWT token
  /api/me:
    get:
      security:
        - bearerAuth: []
      summary: Current user
      responses:
        '200':
          description: User profile
  /api/pcaps/upload:
    post:
      security:
        - bearerAuth: []
      summary: Upload PCAP
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                pcap:
                  type: string
                  format: binary
      responses:
        '200':
          description: Job created
  /api/pcaps:
    get:
      security:
        - bearerAuth: []
      summary: List PCAPs
      responses:
        '200':
          description: List PCAPs
  /api/pcaps/{id}:
    get:
      security:
        - bearerAuth: []
      summary: Get PCAP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: PCAP detail
    delete:
      security:
        - bearerAuth: []
      summary: Delete PCAP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
  /api/pcaps/{id}/jobs:
    get:
      security:
        - bearerAuth: []
      summary: List jobs for PCAP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Job list
  /api/jobs/{id}:
    get:
      security:
        - bearerAuth: []
      summary: Get job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Job detail
  /api/pcaps/{id}/flows:
    get:
      security:
        - bearerAuth: []
      summary: List flows
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: client_ip
          in: query
          schema:
            type: string
        - name: server_ip
          in: query
          schema:
            type: string
        - name: port
          in: query
          schema:
            type: integer
        - name: tcp_stream
          in: query
          schema:
            type: integer
        - name: stream
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Flow list
  /api/jobs/{id}/flows:
    get:
      security:
        - bearerAuth: []
      summary: List flows for job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: client_ip
          in: query
          schema:
            type: string
        - name: server_ip
          in: query
          schema:
            type: string
        - name: port
          in: query
          schema:
            type: integer
        - name: tcp_stream
          in: query
          schema:
            type: integer
        - name: stream
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Flow list for job
  /api/flows/{id}:
    get:
      security:
        - bearerAuth: []
      summary: Get flow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Flow detail
  /api/pcaps/{id}/issues:
    get:
      security:
        - bearerAuth: []
      summary: List issues
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: severity
          in: query
          schema:
            type: integer
        - name: issue_type
          in: query
          schema:
            type: string
        - name: job_id
          in: query
          schema:
            type: integer
        - name: flow_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Issue list (latest job by default)
  /api/jobs/{id}/issues:
    get:
      security:
        - bearerAuth: []
      summary: List issues for job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: severity
          in: query
          schema:
            type: integer
        - name: issue_type
          in: query
          schema:
            type: string
        - name: flow_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Issue list for job
  /api/pcaps/{id}/stats:
    get:
      security:
        - bearerAuth: []
      summary: Get PCAP stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Stats payload
  /api/pcaps/{id}/summary:
    get:
      security:
        - bearerAuth: []
      summary: Get summary cards
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Summary counts
  /api/issues/{id}/explain:
    post:
      security:
        - bearerAuth: []
      summary: Explain issue with AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Explanation
  /api/issues/{id}:
    get:
      security:
        - bearerAuth: []
      summary: Get issue detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Issue detail with evidence
  /api/flows/{id}/cert-inspect:
    post:
      security:
        - bearerAuth: []
      summary: Active certificate inspection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Certificate report